backend/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
IPy==1.0
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-telegram-bot==12.6.1
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
